+++ Riot/Modules/Common/SegmentedViewController/SegmentedViewController.m
@@ -212,6 +212,8 @@
 {
     [super viewWillAppear:animated];
 
+    [self userInterfaceThemeDidChange];
+
     if (_selectedViewController)
     {
         // Make iOS invoke child viewWillAppear