+++ Riot/Modules/KeyVerification/Common/Verify/Scanning/KeyVerificationVerifyByScanningViewModel.swift
@@ -64,8 +64,6 @@ final class KeyVerificationVerifyByScanningViewModel: KeyVerificationVerifyBySca
             self.scannedQRCode(payloadData: payloadData)
         case .cannotScan:
             self.startSASVerification()
-        case .acknowledgeOtherScannedMyCode(let acknowledgeOtherScannedMyCode):
-            self.acknowledgeOtherScannedMyCode(acknowledgeOtherScannedMyCode)
         case .cancel:
             self.cancel()
         case .acknowledgeMyUserScannedOtherCode:
@@ -111,6 +109,7 @@ final class KeyVerificationVerifyByScanningViewModel: KeyVerificationVerifyBySca
     private func cancel() {
         self.cancelQRCodeTransaction()
         self.keyVerificationRequest.cancel(with: MXTransactionCancelCode.user(), success: nil, failure: nil)
+        self.unregisterTransactionDidStateChangeNotification()
         self.coordinatorDelegate?.keyVerificationVerifyByScanningViewModelDidCancel(self)
     }
     
@@ -120,6 +119,7 @@ final class KeyVerificationVerifyByScanningViewModel: KeyVerificationVerifyBySca
         }
         
         transaction.cancel(with: MXTransactionCancelCode.user())
+        self.removePendingQRCodeTransaction()
     }
     
     private func update(viewState: KeyVerificationVerifyByScanningViewState) {
@@ -145,18 +145,9 @@ final class KeyVerificationVerifyByScanningViewModel: KeyVerificationVerifyBySca
             return
         }
         
-        self.update(viewState: .loading)
+        self.unregisterTransactionDidStateChangeNotification()
         self.coordinatorDelegate?.keyVerificationVerifyByScanningViewModel(self, didScanOtherQRCodeData: scannedQRCodeData, withTransaction: qrCodeTransaction)
-    }
-    
-    private func acknowledgeOtherScannedMyCode(_ acknowledgeOtherScannedMyCode: Bool) {
-        guard let qrCodeTransaction = self.qrCodeTransaction else {
-            return
-        }
-        
-        self.update(viewState: .loading)
-        self.coordinatorDelegate?.keyVerificationVerifyByScanningViewModel(self, qrCodeDidScannedByOtherWithTransaction: qrCodeTransaction)
-    }
+    }    
     
     private func removePendingQRCodeTransaction() {
         guard let qrCodeTransaction = self.qrCodeTransaction else {
@@ -250,7 +241,8 @@ final class KeyVerificationVerifyByScanningViewModel: KeyVerificationVerifyBySca
             self.unregisterTransactionDidStateChangeNotification()
             self.coordinatorDelegate?.keyVerificationVerifyByScanningViewModelDidCancel(self)
         case .qrScannedByOther:
-            self.update(viewState: .otherUserScannedMyCode)
+            self.unregisterTransactionDidStateChangeNotification()
+            self.coordinatorDelegate?.keyVerificationVerifyByScanningViewModel(self, qrCodeDidScannedByOtherWithTransaction: transaction)
         case .cancelled:
             guard let reason = transaction.reasonCancelCode else {
                 return