+++ Riot/Modules/Room/DataSources/RoomDataSource.m
@@ -802,7 +802,7 @@
     Widget *jitsiWidget;
 
     // Note: Manage only one jitsi widget at a time for the moment
-    jitsiWidget = [[WidgetManager sharedManager] widgetsOfTypes:@[kWidgetTypeJitsi] inRoom:self.room withRoomState:self.roomState].firstObject;
+    jitsiWidget = [[WidgetManager sharedManager] widgetsOfTypes:@[kWidgetTypeJitsiV1, kWidgetTypeJitsiV2] inRoom:self.room withRoomState:self.roomState].firstObject;
 
     return jitsiWidget;
 }