+++ Riot/Modules/Room/Members/RoomParticipantsViewController.h
@@ -73,7 +73,6 @@
 @property (weak, nonatomic) IBOutlet UIView *searchBarHeaderBorder;
 
 @property (weak, nonatomic) IBOutlet NSLayoutConstraint *searchBarTopConstraint;
-@property (weak, nonatomic) IBOutlet NSLayoutConstraint *tableViewBottomConstraint;
 
 /**
  A matrix room (nil by default).